seprate common test structs
