+++ bedrock/tests/test_smart_account_transfer.rs
@@ -2,12 +2,10 @@ use std::{str::FromStr, sync::Arc};
 
 use alloy::{
     network::Ethereum,
-    node_bindings::AnvilInstance,
-    primitives::{address, keccak256, Address, FixedBytes, Log, U256},
+    primitives::{address, keccak256, Address, U256},
     providers::{ext::AnvilApi, Provider, ProviderBuilder},
     signers::local::PrivateKeySigner,
-    sol,
-    sol_types::{SolCall, SolEvent},
+    sol_types::SolCall,
 };
 
 use bedrock::{
@@ -15,200 +13,20 @@ use bedrock::{
         http_client::{
             set_http_client, AuthenticatedHttpClient, HttpError, HttpMethod,
         },
-        Network, PrimitiveError,
+        Network,
     },
-    smart_account::{
-        EncodedSafeOpStruct, ISafe4337Module, SafeSmartAccount, UserOperation,
-        ENTRYPOINT_4337,
-    },
+    smart_account::{ISafe4337Module, SafeSmartAccount, ENTRYPOINT_4337},
 };
 
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 
-// ------------------ On-chain interfaces used in the test ------------------
+mod common;
+use common::{deploy_safe, setup_anvil, IEntryPoint, IERC20};
 
-sol!(
-    #[allow(missing_docs)]
-    #[sol(rpc)]
-    interface ISafeProxyFactory {
-        event ProxyCreation(address indexed proxy, address singleton);
+// PackedUserOperation and interfaces are imported from common
 
-        function createProxyWithNonce(
-            address _singleton,
-            bytes memory initializer,
-            uint256 saltNonce
-        ) external returns (address proxy);
-    }
-);
-
-sol!(
-    /// The `setup` function of the Safe Smart Account. Sets an initial storage of the Safe contract.
-    #[allow(clippy::too_many_arguments)]
-    #[sol(rpc)]
-    interface ISafe {
-        function setup(
-            address[] calldata _owners,
-            uint256 _threshold,
-            address to,
-            bytes calldata data,
-            address fallbackHandler,
-            address paymentToken,
-            uint256 payment,
-            address payable paymentReceiver
-        ) external;
-
-        function enableModules(address[] memory modules) external;
-    }
-);
-
-sol! {
-    /// A gas efficient representation of a `UserOperation` for use with the `EntryPoint` contract.
-    #[sol(rename_all = "camelCase")]
-    struct PackedUserOperation {
-        address sender;
-        uint256 nonce;
-        bytes init_code;
-        bytes call_data;
-        bytes32 account_gas_limits; // 16 bytes verificationGasLimit | 16 bytes callGasLimit
-        uint256 pre_verification_gas;
-        bytes32 gas_fees; // 16 bytes maxPriorityFeePerGas | 16 bytes maxFeePerGas
-        bytes paymaster_and_data;
-        bytes signature;
-    }
-
-    /// Entry Point Contract
-    #[sol(rpc)]
-    interface IEntryPoint {
-        function depositTo(address account) external payable;
-        function handleOps(PackedUserOperation[] calldata ops, address payable beneficiary) external;
-    }
-
-    /// ERC-20 Token
-    #[sol(rpc)]
-    interface IERC20 {
-        function transfer(address to, uint256 amount) external returns (bool);
-        function balanceOf(address account) external view returns (uint256);
-        function approve(address spender, uint256 amount) external returns (bool);
-    }
-}
-
-fn pack_pair(a: &u128, b: &u128) -> FixedBytes<32> {
-    let mut out = [0u8; 32];
-    out[..16].copy_from_slice(&a.to_be_bytes());
-    out[16..].copy_from_slice(&b.to_be_bytes());
-    out.into()
-}
-
-impl TryFrom<&UserOperation> for PackedUserOperation {
-    type Error = PrimitiveError;
-
-    fn try_from(user_op: &UserOperation) -> Result<Self, Self::Error> {
-        Ok(Self {
-            sender: user_op.sender,
-            nonce: user_op.nonce,
-            init_code: user_op.get_init_code(),
-            call_data: user_op.call_data.clone(),
-            account_gas_limits: pack_pair(
-                &user_op.verification_gas_limit,
-                &user_op.call_gas_limit,
-            ),
-            pre_verification_gas: user_op.pre_verification_gas,
-            gas_fees: pack_pair(
-                &user_op.max_priority_fee_per_gas,
-                &user_op.max_fee_per_gas,
-            ),
-            paymaster_and_data: user_op.get_paymaster_and_data(),
-            signature: user_op.signature.clone(),
-        })
-    }
-}
-
-// ------------------ Anvil helper setup ------------------
-
-const SAFE_PROXY_FACTORY_ADDRESS: Address =
-    address!("4e1DCf7AD4e460CfD30791CCC4F9c8a4f820ec67");
-const SAFE_L2_SINGLETON_ADDRESS: Address =
-    address!("29fcB43b46531BcA003ddC8FCB67FFE91900C762");
-const SAFE_4337_MODULE_ADDRESS: Address =
-    address!("75cf11467937ce3F2f357CE24ffc3DBF8fD5c226");
-const SAFE_MODULE_SETUP_ADDRESS: Address =
-    address!("2dd68b007B46fBe91B9A7c3EDa5A7a1063cB5b47");
-
-fn setup_anvil() -> AnvilInstance {
-    dotenvy::dotenv().ok();
-    let rpc_url = std::env::var("WORLDCHAIN_RPC_URL").unwrap_or_else(|_| {
-        // Fallback to a public, no-key RPC if available. If this fails, please set WORLDCHAIN_RPC_URL.
-        // NOTE: Replace with a working public endpoint if this one is unavailable.
-        "https://worldchain-mainnet.g.alchemy.com/v2/demo".to_string()
-    });
-
-    alloy::node_bindings::Anvil::new().fork(rpc_url).spawn()
-}
-
-async fn deploy_safe<P>(
-    provider: &P,
-    owner: Address,
-    deploy_nonce: U256,
-) -> anyhow::Result<Address>
-where
-    P: Provider<Ethereum>,
-{
-    provider
-        .anvil_set_balance(owner, U256::from(1e19 as u64))
-        .await
-        .unwrap();
-
-    let proxy_factory = ISafeProxyFactory::new(SAFE_PROXY_FACTORY_ADDRESS, provider);
-
-    let setup_data = ISafe::setupCall {
-        _owners: vec![owner],
-        _threshold: U256::from(1),
-        to: SAFE_MODULE_SETUP_ADDRESS,
-        data: ISafe::enableModulesCall {
-            modules: vec![SAFE_4337_MODULE_ADDRESS],
-        }
-        .abi_encode()
-        .into(),
-        fallbackHandler: SAFE_4337_MODULE_ADDRESS,
-        paymentToken: Address::ZERO,
-        payment: U256::ZERO,
-        paymentReceiver: Address::ZERO,
-    }
-    .abi_encode();
-
-    let deploy_tx = proxy_factory
-        .createProxyWithNonce(
-            SAFE_L2_SINGLETON_ADDRESS,
-            setup_data.into(),
-            deploy_nonce,
-        )
-        .from(owner)
-        .send()
-        .await
-        .expect("Failed to send createProxyWithNonce transaction");
-
-    let receipt = deploy_tx
-        .get_receipt()
-        .await
-        .expect("Failed to get transaction receipt");
-
-    // Find ProxyCreation event and extract proxy address
-    let proxy_creation_event = receipt
-        .inner
-        .logs()
-        .iter()
-        .find_map(|log| {
-            let raw_log = Log {
-                address: log.address(),
-                data: log.data().clone(),
-            };
-            ISafeProxyFactory::ProxyCreation::decode_log(&raw_log).ok()
-        })
-        .expect("ProxyCreation event not found");
-
-    Ok(proxy_creation_event.proxy)
-}
+// setup_anvil and deploy_safe are provided by common
 
 // ------------------ Mock HTTP client that actually executes the op on Anvil ------------------
 
@@ -218,12 +36,11 @@ where
     P: Provider<Ethereum> + Clone + Send + Sync + 'static,
 {
     provider: P,
-    beneficiary: Address,
 }
 
 #[derive(Deserialize)]
 struct JsonRpcRequestLite {
-    jsonrpc: String,
+    _jsonrpc: String,
     id: Value,
     method: String,
     params: Value,
@@ -264,12 +81,24 @@ where
             message: "missing body".into(),
         })?;
 
-        let req: JsonRpcRequestLite =
+        let root: serde_json::Value =
             serde_json::from_slice(&body).map_err(|_| HttpError::Generic {
                 message: "invalid json".into(),
             })?;
 
-        match req.method.as_str() {
+        let method =
+            root.get("method")
+                .and_then(|m| m.as_str())
+                .ok_or(HttpError::Generic {
+                    message: "invalid json".into(),
+                })?;
+        let id = root.get("id").cloned().unwrap_or(serde_json::Value::Null);
+        let params = root
+            .get("params")
+            .cloned()
+            .unwrap_or(serde_json::Value::Null);
+
+        match method {
             // Respond with minimal, sane gas values and no paymaster
             "wa_sponsorUserOperation" => {
                 let result = SponsorUserOperationResponseLite {
@@ -285,17 +114,17 @@ where
                 };
                 let resp = json!({
                     "jsonrpc": "2.0",
-                    "id": req.id,
+                    "id": id,
                     "result": result,
                 });
                 Ok(serde_json::to_vec(&resp).unwrap())
             }
             // Execute the inner call directly through the Safe 4337 Module (no sponsorship path)
             "eth_sendUserOperation" => {
-                let params = req.params.as_array().ok_or(HttpError::Generic {
+                let params = params.as_array().ok_or(HttpError::Generic {
                     message: "invalid params".into(),
                 })?;
-                let user_op_val = params.get(0).ok_or(HttpError::Generic {
+                let user_op_val = params.first().ok_or(HttpError::Generic {
                     message: "missing userOp param".into(),
                 })?;
                 let _entry_point_str = params.get(1).and_then(|v| v.as_str()).ok_or(
@@ -304,47 +133,38 @@ where
                     },
                 )?;
 
-                // Fill in optional fields if missing to satisfy deserialization
-                let mut op_obj =
-                    user_op_val.as_object().cloned().ok_or(HttpError::Generic {
-                        message: "userOp param must be an object".into(),
+                // Validate presence of required fields (no defaulting). Throw if invalid.
+                let op_obj = user_op_val.as_object().ok_or(HttpError::Generic {
+                    message: "userOp param must be an object".into(),
+                })?;
+
+                let sender_str = op_obj.get("sender").and_then(|v| v.as_str()).ok_or(
+                    HttpError::Generic {
+                        message: "missing or invalid sender".into(),
+                    },
+                )?;
+                let call_data_str = op_obj
+                    .get("callData")
+                    .and_then(|v| v.as_str())
+                    .ok_or(HttpError::Generic {
+                        message: "missing or invalid callData".into(),
                     })?;
 
-                op_obj.entry("factory").or_insert(Value::String(
-                    "0x0000000000000000000000000000000000000000".into(),
-                ));
-                op_obj
-                    .entry("factoryData")
-                    .or_insert(Value::String("0x".into()));
-                op_obj.entry("paymaster").or_insert(Value::String(
-                    "0x0000000000000000000000000000000000000000".into(),
-                ));
-                op_obj
-                    .entry("paymasterVerificationGasLimit")
-                    .or_insert(Value::Number(0u64.into()));
-                op_obj
-                    .entry("paymasterPostOpGasLimit")
-                    .or_insert(Value::Number(0u64.into()));
-                op_obj
-                    .entry("paymasterData")
-                    .or_insert(Value::String("0x".into()));
+                let sender =
+                    Address::from_str(sender_str).map_err(|_| HttpError::Generic {
+                        message: "invalid sender".into(),
+                    })?;
 
-                let fixed_user_op_val = Value::Object(op_obj);
-
-                let user_op: UserOperation = serde_json::from_value(fixed_user_op_val)
-                    .map_err(|e| {
-                        let payload_str =
-                            serde_json::to_string(user_op_val).unwrap_or_default();
-                        HttpError::Generic {
-                            message: format!(
-                                "invalid userOp: {e}; payload: {payload_str}"
-                            ),
-                        }
+                let call_data_hex =
+                    call_data_str.strip_prefix("0x").unwrap_or(call_data_str);
+                let call_data_bytes =
+                    hex::decode(call_data_hex).map_err(|_| HttpError::Generic {
+                        message: "invalid callData".into(),
                     })?;
 
                 // Decode the module callData and simulate a plain ERC20 transfer by updating storage
                 let module_call =
-                    ISafe4337Module::executeUserOpCall::abi_decode(&user_op.call_data)
+                    ISafe4337Module::executeUserOpCall::abi_decode(&call_data_bytes)
                         .map_err(|e| HttpError::Generic {
                             message: format!(
                                 "decode executeUserOp callData failed: {e}"
@@ -368,7 +188,7 @@ where
                 // Read current balances
                 let erc20 = IERC20::new(token_address, &self.provider);
                 let sender_balance =
-                    erc20.balanceOf(user_op.sender).call().await.map_err(|e| {
+                    erc20.balanceOf(sender).call().await.map_err(|e| {
                         HttpError::Generic {
                             message: format!("read sender balance failed: {e}"),
                         }
@@ -388,7 +208,7 @@ where
                     U256::from_be_bytes(slot_hash.into())
                 };
 
-                let sender_slot = calc_slot(user_op.sender);
+                let sender_slot = calc_slot(sender);
                 let recipient_slot = calc_slot(recipient);
 
                 // Update balances
@@ -412,18 +232,16 @@ where
                         message: format!("set recipient storage failed: {e}"),
                     })?;
 
-                // Return a deterministic 32-byte hash derived from the encoded safe op
-                let encoded =
-                    EncodedSafeOpStruct::try_from(&user_op).map_err(|_| {
-                        HttpError::Generic {
-                            message: "encode failed".into(),
-                        }
-                    })?;
-                let user_op_hash: FixedBytes<32> = encoded.into_transaction_hash();
+                // Return a deterministic pseudo userOpHash from sender+callData
+                let mut preimage =
+                    Vec::with_capacity(sender.as_slice().len() + call_data_bytes.len());
+                preimage.extend_from_slice(sender.as_slice());
+                preimage.extend_from_slice(&call_data_bytes);
+                let user_op_hash = keccak256(preimage);
 
                 let resp = json!({
                     "jsonrpc": "2.0",
-                    "id": req.id,
+                    "id": id,
                     "result": format!("0x{}", hex::encode(user_op_hash)),
                 });
                 Ok(serde_json::to_vec(&resp).unwrap())
@@ -489,7 +307,6 @@ async fn test_transaction_transfer_full_flow_executes_user_operation(
     // 7) Install mocked HTTP client that routes calls to Anvil
     let client = AnvilBackedHttpClient {
         provider: provider.clone(),
-        beneficiary: owner,
     };
     let _ = set_http_client(Arc::new(client));
